@@ -1,6 +1,6 @@
 {
 	"name": "ansi-styles",
-	"version": "4.3.0",
+	"version": "5.2.0",
 	"description": "ANSI escape codes for styling strings in the terminal",
 	"license": "MIT",
 	"repository": "chalk/ansi-styles",
@@ -8,10 +8,10 @@
 	"author": {
 		"name": "Sindre Sorhus",
 		"email": "sindresorhus@gmail.com",
-		"url": "sindresorhus.com"
+		"url": "https://sindresorhus.com"
 	},
 	"engines": {
-		"node": ">=8"
+		"node": ">=10"
 	},
 	"scripts": {
 		"test": "xo && ava && tsd",
@@ -43,14 +43,10 @@
 		"command-line",
 		"text"
 	],
-	"dependencies": {
-		"color-convert": "^2.0.1"
-	},
 	"devDependencies": {
-		"@types/color-convert": "^1.9.0",
-		"ava": "^2.3.0",
+		"ava": "^2.4.0",
 		"svg-term-cli": "^2.1.1",
-		"tsd": "^0.11.0",
-		"xo": "^0.25.3"
+		"tsd": "^0.14.0",
+		"xo": "^0.37.1"
 	}
 }
