@@ -31,5 +31,5 @@ class Circle extends Shape {
 module.exports= {
     Shape,
     Triangle,
-    Circle,
+    Circle
 };
